refactor: lookup dialogs
